Add instructions for rendering AnimatedSprite
Since there is not an example spritesheet .png here and there are a lot of parameters to potentially get wrong in the atlasData, it can be a little hard to debug why the sprite is not animating for a new person like myself.

I'd like to propose adding a few lines to loop the animated sprite and add it to the stage, since other examples seem to include a line to add the object to the stage.

Also, if the animationSpeed line is not needed generally I can remove it. I ran into issues using the default of `1` but perhaps that's a separate issue.

Thanks for maintaining a fantastic rendering library!